+++ Logitude/Unifreight.Data/AmitalModel/EntityPOCOs/AmitalDataModel.CCUTRANSPVAL.cs
@@ -78,7 +78,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
         /// There are no comments for TRANSPVALFC in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> TRANSPVALFC
+        public virtual global::System.Nullable<decimal> TRANSPVALFC
         {
             get
             {
@@ -93,7 +93,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _TRANSPVALFC;
+        private global::System.Nullable<decimal> _TRANSPVALFC;
 
     
         /// <summary>
@@ -122,7 +122,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
         /// There are no comments for TRANSPVAL in the schema.
         /// </summary>
         [System.Runtime.Serialization.DataMember]
-        public virtual global::System.Nullable<long> TRANSPVAL
+        public virtual global::System.Nullable<decimal> TRANSPVAL
         {
             get
             {
@@ -137,7 +137,7 @@ namespace Unifreight.Data.AmitalModel.EntityPOCOs
                 }
             }
         }
-        private global::System.Nullable<long> _TRANSPVAL;
+        private global::System.Nullable<decimal> _TRANSPVAL;
 
     
         /// <summary>
